refactor: remove console.log messages and split code to modules
