+++ src/control/mouse_position.ts
@@ -1,0 +1,6 @@
+import robot from 'robotjs';
+
+export const getMousePosition = () => {
+  const position = robot.getMousePos();
+  return position;
+};
