+++ src/control/screen_image.ts
